+++ packages/graph-viz/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cognidocs/graph-viz
 
+## 1.0.1
+
+### Patch Changes
+
+- Fixed markdown rendering for generated content
+
 ## 1.0.0
 
 ### Major Changes