+++ packages/plugin-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cognidocs/plugin-core
 
+## 1.0.1
+
+### Patch Changes
+
+- Fixed markdown rendering for generated content
+- Updated dependencies
+  - @cognidocs/utils@1.0.1
+
 ## 1.0.0
 
 ### Major Changes